requirements.txt
# Simplified requirements based on the Milvus Quickstart Guide

# Core Vector DB libraries
pymilvus
milvus-lite

# Machine Learning & Embeddings
torch
sentence-transformers
scikit-learn
numpy
neo4j
semantic-chunking 
langchain-experimental
# sentence-transformers 
# cross-encoder

# Document Processing
pypdf
langchain-text-splitters

# LLM Integration
google-generativeai